+++ lib/src/statements/sql_statement.dart
@@ -1,4 +1,4 @@
-import 'package:postgres_builder/src/processed_sql.dart';
+import 'package:postgres_builder/postgres_builder.dart';
 
 // ignore: one_member_abstracts
 abstract class SqlStatement {
@@ -6,4 +6,9 @@ abstract class SqlStatement {
   ProcessedSql toSql();
 }
 
-abstract class FilterStatement implements SqlStatement {}
+abstract class FilterStatement implements SqlStatement {
+  const FilterStatement();
+
+  FilterStatement operator &(FilterStatement other) => And([this, other]);
+  FilterStatement operator |(FilterStatement other) => Or([this, other]);
+}